+++ src/DependencyInjection/NelmioApiDocExtension.php
@@ -21,6 +21,7 @@ use Nelmio\ApiDocBundle\Describer\RouteDescriber;
 use Nelmio\ApiDocBundle\ModelDescriber\BazingaHateoasModelDescriber;
 use Nelmio\ApiDocBundle\ModelDescriber\JMSModelDescriber;
 use Nelmio\ApiDocBundle\ModelDescriber\ModelDescriberInterface;
+use Nelmio\ApiDocBundle\OpenApiGenerator;
 use Nelmio\ApiDocBundle\Processor\MapQueryStringProcessor;
 use Nelmio\ApiDocBundle\Processor\MapRequestPayloadProcessor;
 use Nelmio\ApiDocBundle\RouteDescriber\RouteArgumentDescriber;
@@ -74,7 +75,7 @@ final class NelmioApiDocExtension extends Extension implements PrependExtensionI
         $container->setParameter('nelmio_api_doc.use_validation_groups', $config['use_validation_groups']);
 
         // Register the OpenAPI Generator as a service.
-        $container->register('nelmio_api_doc.open_api.generator', Generator::class)
+        $container->register('nelmio_api_doc.open_api.generator', OpenApiGenerator::class)
             ->setPublic(false);
 
         $cachePool = $config['cache']['pool'] ?? null;
@@ -160,6 +161,11 @@ final class NelmioApiDocExtension extends Extension implements PrependExtensionI
                 array_keys($config['areas']),
                 array_map(function ($area) { return new Reference(\sprintf('nelmio_api_doc.generator.%s', $area)); }, array_keys($config['areas']))
             ));
+
+        if (true === $config['type_info']) {
+            $container->getDefinition('nelmio_api_doc.model_describers.object')
+                ->setArgument(2, new Reference('nelmio_api_doc.type_describer.chain'));
+        }
 
         $container->getDefinition('nelmio_api_doc.model_describers.object')
             ->setArgument(2, $config['media_types']);