+++ src/PropertyDescriber/ObjectPropertyDescriber.php
@@ -36,6 +36,13 @@ final class ObjectPropertyDescriber implements PropertyDescriberInterface, Model
             $types[0]->getCollectionValueTypes()[0] ?? null,
         ); // ignore nullable field
 
+        if (null === $types[0]->getClassName()) {
+            $property->type = 'object';
+            $property->additionalProperties = true;
+
+            return;
+        }
+
         if ($types[0]->isNullable()) {
             $weakContext = Util::createWeakContext($property->_context);
             $schemas = [new OA\Schema(['ref' => $this->modelRegistry->register(new Model($type, serializationContext: $context)), '_context' => $weakContext])];