+++ tests/Functional/ModelDescriber/Fixtures/DateTimeClass.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * This file is part of the NelmioApiDocBundle package.
+ *
+ * (c) Nelmio
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Nelmio\ApiDocBundle\Tests\Functional\ModelDescriber\Fixtures;
+
+class DateTimeClass
+{
+    public \DateTime $dateTime;
+    public \DateTimeImmutable $dateTimeImmutable;
+    public \DateTimeInterface $dateTimeInterface;
+
+    public ?\DateTime $nullableDateTime;
+}