+++ tests/Functional/ModelDescriber/Fixtures/TypeInfo/ArrayMixedKeys.php
@@ -1,0 +1,20 @@
+<?php
+
+/*
+ * This file is part of the NelmioApiDocBundle package.
+ *
+ * (c) Nelmio
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Nelmio\ApiDocBundle\Tests\Functional\ModelDescriber\Fixtures\TypeInfo;
+
+class ArrayMixedKeys
+{
+    /**
+     * @var array<int | string, int>
+     */
+    public array $mixedArray;
+}