+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG
 
+## 4.35.0
+* Added support for the symfony/type-info component
+```yaml
+nelmio_api_doc:
+  type_info: true
+```
+
+## 4.34.0
+* Changed minimum Symfony version for 7.x from 7.0 to 7.1
+
+## 4.33.6
+* Fixed Symfony 7.2 deprecation of tagged arguments
+
+## 4.33.5
+* Added new optional parameter `$context` to` PropertyDescriberInterface::supports()`
+
 ## 4.33.4
 * Deprecated `null` type from `$options` in `Nelmio\ApiDocBundle\Attribute\Model::__construct()`. Pass an empty array (`[]`) instead.
 * Deprecated `null` type from `$options` in `NNelmio\ApiDocBundle\Attribute\Model::__construct()`. Pass an empty array (`[]`) instead.